affine refactor
